+++ src/main/java/com/wondumall/Controller/QuestionController.java
@@ -82,11 +82,10 @@ public class QuestionController {
 		}
 		
 		//관리자가 아니거나 작성자가 아닐 경우 접근 에러
-		boolean adminCheck = !myUserDetails.getAuthorities().stream().anyMatch(a -> a.getAuthority().equals("ROLE_ADMIN")); 
-		
-		if(adminCheck && !detail.getU_nickname().equals(myUserDetails.getNickname())) {
-	         throw new ResponseStatusException(HttpStatus.FORBIDDEN);
-	      }
+		boolean adminCheck = myUserDetails.getAuthorities().stream().anyMatch(a -> a.getAuthority().equals("ROLE_ADMIN")); 
+		if(!adminCheck && !detail.getU_nickname().equals(myUserDetails.getNickname())){ // 관리자가 아니거나 글 작성자가 아닐 경우
+			throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+		}
 		
 		mv.addObject("detail", detail);
 		mv.addObject("pageNo", pageNo);
@@ -153,9 +152,19 @@ public class QuestionController {
 	
 	@Secured({"ROLE_USER", "ROLE_BUISNESS", "ROLE_ADMIN"})
 	@GetMapping("/questionDelete.do")
-	public void questionDelete(@RequestParam("pageNo") int pageNo, @RequestParam("q_no") int q_no, HttpServletRequest request,
+	public void questionDelete(@RequestParam("pageNo") int pageNo, @RequestParam("q_no") int q_no, @RequestParam("u_nickname") String u_nickname, HttpServletRequest request,
 			HttpServletResponse response, @RequestParam(name="searchColumn", required = false) String searchColumn, 
-			@RequestParam(name="searchValue", required=false) String searchValue) throws Exception {
+			@RequestParam(name="searchValue", required=false) String searchValue,
+			@AuthenticationPrincipal MyUserDetails myUserDetails) throws Exception {
+		//비로그인 회원일 경우 접근 에러
+		if(myUserDetails== null) {
+			throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+		}
+		boolean adminCheck = myUserDetails.getAuthorities().stream().anyMatch(a -> a.getAuthority().equals("ROLE_ADMIN")); 
+		if(!adminCheck && !u_nickname.equals(myUserDetails.getNickname())){ // 관리자가 아니거나 글 작성자가 아닐 경우
+			throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+		}
+		
 		QuestionDTO questionDTO = new QuestionDTO();
 		questionDTO.setQ_no(q_no);
 		int result = questionService.delete(questionDTO);
